Some additions and changes
Added the div that will be where the menu for each different food category will be displayed, made it scrollable with a scroll bar that only appears when the cursors hovers above the div. Added a select in the orderBox div, and changed the arrow with one from online. Added an assets folder