Scrolling Menu Items
Added a box in the bottom of the menuBox that scrolls through the menu items. They darken when hovered over and there are buttons and prices. Changed the size of the menu and order box again. Added the alert & alert-dismissable classes to the ordered items list so that they can be closed - This is a temporary thing